requirements.txt
python-dotenv
livekit-agents[sarvam,openai,silero,turn-detector]~=1.0
livekit-plugins-noise-cancellation~=0.2
torch